Support gate level input

+++ Makefile
@@ -31,8 +31,8 @@ release: run-cmake-release
 	cp -f ./yosys_verific_rs/yosys/install/bin/yosys build/bin/yosys
 	cp -f ./yosys_verific_rs/yosys/install/bin/abc build/bin/abc
 	cp -f ./yosys_verific_rs/yosys/install/bin/de build/bin/de
-	cp -f ./OpenFPGA_RS/openfpga/openfpga build/bin/openfpga
 	@[ -f ./OpenFPGA_RS/vpr/vpr ] && cp -f ./OpenFPGA_RS/vpr/vpr build/bin/vpr || true
+	@[ -f ./OpenFPGA_RS/openfpga/openfpga ] && cp -f ./OpenFPGA_RS/openfpga/openfpga build/bin/openfpga || true
 
 release_no_tcmalloc: run-cmake-release_no_tcmalloc
 	cmake --build build -j $(CPU_CORES)
@@ -42,8 +42,8 @@ debug: run-cmake-debug
 	cp -f ./yosys_verific_rs/yosys/debug-install/bin/yosys dbuild/bin/yosys
 	cp -f ./yosys_verific_rs/yosys/debug-install/bin/abc dbuild/bin/abc
 	cp -f ./yosys_verific_rs/yosys/debug-install/bin/de dbuild/bin/de
-	cp -f ./OpenFPGA_RS/openfpga/openfpga dbuild/bin/openfpga
 	@[ -f ./OpenFPGA_RS/vpr/vpr ] && cp -f ./OpenFPGA_RS/vpr/vpr dbuild/bin/vpr || true
+	@[ -f ./OpenFPGA_RS/openfpga/openfpga ] && cp -f ./OpenFPGA_RS/openfpga/openfpga dbuild/bin/openfpga || true
 
 run-cmake-release:
 	cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$(PREFIX) -DCMAKE_RULE_MESSAGES=$(RULE_MESSAGES) $(ADDITIONAL_CMAKE_OPTIONS) -S . -B build
